Correct the link of libraries wrap.so automatically created.
This cancels the default link flag --as-needed, that is added
on compilers since Natty.
Because of this flag, only part of the libraries were linked
with the library wrap.so (libpython was linked, e.g.).
Especially the library normally attached to wrap.so (any plugin of the
stack of tasks) was not linked, since no explicit call to the library
is made.
This cause an error in the import of symbols in python.
By adding the --no-as-needed flag, the link is re-established.