Tag dependencies includes as system includes.
To avoid importing warnings from external dependencies, the call to
INCLUDE_DIRECTORIES in pkg-config.cmake is now specifying the SYSTEM
keyword to import header directory using the special -isystem flag
when using gcc. These headers will now be considered as "system"
headers and warnings will be discarded for them.

This should allow us to maintain the -Werror flag in as many packages
as possible.