python.cmake: add a PYTHON_BUILD macro.
This macros provides a way to build some source file and store them in
the build directory. Then, the install macro can copy those files to the
adequate destination.

Thus, the compilation can be done without calling the "install" target,
which makes it easier to test local modules. We rely on
ADD_CUSTOM_TARGET for the compilation of each file.