[Win32] Correct pc file when compiling with Visual.
More specifically, the CMAKE_CONFIGURATION_TYPE variable
is always defined when generating the project for Visual Studio
(it equals Release,RelWithDebInfo,Debug...).
With the former version of the code, if one run cmake with the
option CMAKE_BUILD_TYPE activated, the library specified in the .pc
file would always link with the debug version of the lib.