add BINDINGS_ONLY option
For a package which has python bindings, it is convinient to build on
the one hand only the main binary package, and on the other hand several
binary packages with the bindings for different python versions.

For this other packages to be able to co-exist on a system (and ease
transition from Python 2 to Python 3), there must not be any path
conflict, ie. everything must be either in a PYTHON_SITELIB or with a
SOABI depending on the python version.

The automatically generated
- lib/pkgconfig/<project>.pc
- include/<project>/config.hh
- include/<project>/deprecated.hh
- include/<project>/warning.hh
files are preventing this to work.

This PR adds a BINDINGS_ONLY option allows to avoid the installation of
those files.