update GNUInstallDirs.cmake
This is especially useful to stop having to override
CMAKE_INSTALL_LIBDIR, because we get a lib64 default everywhere, even if
nobody has a 32 bits system anymore.

Another option would be to stop shipping this file, which is provided
by CMake >= 2.8.5, but the commit message from 884e3bd4 lacks explicity
about the reasons of the inclusion of this file.