[find-external] fix CppAD file names
ref https://github.com/stack-of-tasks/pinocchio/pull/1244

As we define lowercase variables (eg. `cppad_FOUND`), we should use
lowercase dependency in `ADD_PROJECT_DEPENDENCY(cppad)`. And for this to
work, the file name must be lowercase `Findcppad.cmake`.

Otherwise, we could define `CppAD_FOUND`, and use
`ADD_PROJECT_DEPENDENCY(CppAD)`, but as we already have some uses of
`cppad_SOMETHING` in the wild, it's too late.

We could provide support for both, if anyone requires it.