set a minimal C++ standard
This commit defines a `SET_MINIMAL_CXX_STANDARD()` macro.

This allows to automatically set CMAKE_CXX_STANDARD if it is not already
manually defined somewhere else (on an equal or greater version), and if
we know that we require a standard version higher than the default one
of the current compiler.

This can avoid to put `set(CMAKE_CXX_STANDARD 11)` in CMakeLists.txt
because C++11 features are required, even if more and more compilers
will default to C++14 and more.

Also, when a package doesn't actually require any particular standard,
but one of its dependency has been compiled with C++11, it's important
to also compile in C++11. For this to work, we just need to call this
macro from the generated *Config.cmake.