avoid multiple findpython() calls
This add a guard in exported modules to avoid multiple ones calling
successive findpython(), which can fail (for a yet unknown reason).

This is required when package A and B are independant and use and export
findpython(), and package C depends on both.

eg. hpp-gepetto-viewer configuration fails, because its
`add_project_dependency("hpp-corbaserver")` calls findpython(), and then
`add_project_dependency("gepetto-viewer")` calls it again, and fail.