Fix logging finalization as a subproject (#590)
* Fix logging finalization as a subproject

Previously, including e.g. ProxSuite as a subproject of a CMake project
caused the call to logging_finalize to have an empty
JRL_CMAKEMODULE_LOGGING_FILENAME, which caused files named `##
------------ ##\n` to be created in the project source directory with
contents "the rest of the stuff supposed to be written to
JRL_CMAKEMODULE_LOGGING_FILENAME".  This seems to fix that for my use
case, but I'm not sure if it's generally correct.

Topic: logging-finalize-filename

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>